Move repository classes into separate folder
