+++ __tests__/persistence/ViewRepository.test.js
@@ -1,4 +1,4 @@
-const ViewRepository = require('../../src/persistence/ViewRepository')
+const ViewRepository = require('../../src/persistence/repositories/ViewRepository')
 
 describe('The ViewRepository', () => {
   it('should store a new object', () => {
