+++ src/persistence/repositories/ContentSlotOptionRepository.js
@@ -1,4 +1,4 @@
-const NotFoundError = require('../errors/NotFoundError')
+const NotFoundError = require('../../errors/NotFoundError')
 
 class ContentSlotOptionRepository {
   constructor () {
